Trait sc_consensus_subspace_rpc::SubspaceRpcApiServer

pub trait SubspaceRpcApiServer: Sized + Send + Sync + 'static {
    fn get_farmer_metadata(&self) -> RpcResult<FarmerMetadata>;
    fn submit_solution_response(
        &self, 
        solution_response: SolutionResponse
    ) -> RpcResult<()>;
    fn submit_reward_signature(
        &self, 
        reward_signature: RewardSignatureResponse
    ) -> RpcResult<()>;
    fn acknowledge_archived_segment<'life0, 'async_trait>(
        &'life0 self, 
        segment_index: u64
    ) -> Pin<Box<dyn Future<Output = RpcResult<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn subscribe_slot_info(&self, subscription_sink: PendingSubscription);
    fn subscribe_reward_signing(&self, subscription_sink: PendingSubscription);
    fn subscribe_archived_segment(&self, subscription_sink: PendingSubscription);

    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the SubspaceRpcApi RPC API.

Required Methods

fn get_farmer_metadata(&self) -> RpcResult<FarmerMetadata>

Ger metadata necessary for farmer operation

fn submit_solution_response(
&self,
solution_response: SolutionResponse
) -> RpcResult<()>

fn submit_reward_signature(
&self,
reward_signature: RewardSignatureResponse
) -> RpcResult<()>

fn acknowledge_archived_segment<'life0, 'async_trait>(
&'life0 self,
segment_index: u64
) -> Pin<Box<dyn Future<Output = RpcResult<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

fn subscribe_slot_info(&self, subscription_sink: PendingSubscription)

Slot info subscription

fn subscribe_reward_signing(&self, subscription_sink: PendingSubscription)

Sign block subscription

fn subscribe_archived_segment(&self, subscription_sink: PendingSubscription)

Archived segment subscription

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.